+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Tapestry currently operates a one month release cycle between releases, see the 
 ### Added
 - #105 Added workspace scaffolding
 - #241 Added optional max argument to permalink category template tag 
+- #246 Added getEnvironment view helper
 
 ## [1.0.10] - 2017-08-01
 ### Fixed